Created mand1415 and kore1280
Created workbooks Mandarin_China_mand1415 and Korean_kore1280 and filled them in to the best of my ability. Added medadata to languages and sources to Zotero and sources.bib. Within hours contracted I could not find some sources for words I know to be in use; in these cases "word" and "comment" are filled and "source" has been left blank.